Refactor RestartFailedTasks into separate component
- Created new RestartFailedTasks component for better code organization
- Moved all failed task collection and restart logic to the new component
- Simplified RestartPatch component, which now just orchestrates the UI
- Exported RestartFailedTasks from PatchActionButtons index
- Improved maintainability and reusability
