+++ apps/spruce/src/components/PatchActionButtons/RestartPatch.tsx
@@ -1,23 +1,10 @@
 import { useState } from "react";
-import { useMutation, useQuery } from "@apollo/client/react";
 import { Button, Size as ButtonSize } from "@leafygreen-ui/button";
-import { MenuItem } from "@leafygreen-ui/menu";
 import { SplitButton } from "@leafygreen-ui/split-button";
 import { Tooltip, Align, Justify } from "@leafygreen-ui/tooltip";
-import { useToastContext } from "@evg-ui/lib/context/toast";
-import { TaskStatus } from "@evg-ui/lib/types/task";
-import { useVersionAnalytics } from "analytics";
 import { DropdownItem } from "components/ButtonDropdown";
 import { VersionRestartModal } from "components/VersionRestartModal";
-import { finishedTaskStatuses } from "constants/task";
-import {
-  BuildVariantsWithChildrenQuery,
-  BuildVariantsWithChildrenQueryVariables,
-  RestartVersionsMutation,
-  RestartVersionsMutationVariables,
-} from "gql/generated/types";
-import { RESTART_VERSIONS } from "gql/mutations";
-import { BUILD_VARIANTS_WITH_CHILDREN } from "gql/queries";
+import { RestartFailedTasks } from "./RestartFailedTasks";
 
 interface RestartPatchProps {
   disabled?: boolean;
@@ -42,112 +29,11 @@ export const RestartPatch: React.FC<RestartPatchProps> = ({
       ? visibilityControl
       : fallbackVisibilityControl;
 
-  const dispatchToast = useToastContext();
-  const { sendEvent } = useVersionAnalytics(patchId);
-
   const onClick = () => setIsVisible(!isVisible);
 
   const message = isMergeQueuePatch
     ? "GitHub merge queue patches cannot be restarted."
     : "This patch cannot be restarted.";
-
-  const [restartVersions, { loading: mutationLoading }] = useMutation<
-    RestartVersionsMutation,
-    RestartVersionsMutationVariables
-  >(RESTART_VERSIONS, {
-    onCompleted: () => {
-      dispatchToast.success(`Successfully restarted tasks!`);
-    },
-    onError: (err) => {
-      dispatchToast.error(`Error while restarting tasks: '${err.message}'`);
-    },
-    refetchQueries,
-  });
-
-  const { data, loading: queryLoading } = useQuery<
-    BuildVariantsWithChildrenQuery,
-    BuildVariantsWithChildrenQueryVariables
-  >(BUILD_VARIANTS_WITH_CHILDREN, {
-    variables: {
-      id: patchId,
-      statuses: [...finishedTaskStatuses, TaskStatus.Aborted],
-    },
-    skip: !isButton,
-  });
-
-  const handleRestartFailedTasks = () => {
-    const { version } = data || {};
-    const { buildVariants, childVersions } = version || {};
-
-    const collectFailedTasks = (
-      variants: BuildVariantsWithChildrenQuery["version"]["buildVariants"],
-    ) => {
-      const taskIds: string[] = [];
-      variants?.forEach((bv) => {
-        bv.tasks?.forEach((task) => {
-          if (
-            task.displayStatus === TaskStatus.Failed ||
-            task.displayStatus === TaskStatus.TaskTimedOut ||
-            task.displayStatus === TaskStatus.TestTimedOut ||
-            task.displayStatus === TaskStatus.KnownIssue ||
-            task.displayStatus === TaskStatus.SetupFailed ||
-            task.displayStatus === TaskStatus.SystemFailed ||
-            task.displayStatus === TaskStatus.SystemTimedOut ||
-            task.displayStatus === TaskStatus.SystemUnresponsive
-          ) {
-            taskIds.push(task.id);
-          }
-        });
-      });
-      return taskIds;
-    };
-
-    const versionsToRestart = [];
-
-    // Collect failed tasks from main version
-    const mainVersionTaskIds = collectFailedTasks(buildVariants);
-    if (mainVersionTaskIds.length > 0) {
-      versionsToRestart.push({
-        versionId: patchId,
-        taskIds: mainVersionTaskIds,
-      });
-    }
-
-    // Collect failed tasks from child versions
-    childVersions?.forEach((childVersion) => {
-      const childTaskIds = collectFailedTasks(childVersion.buildVariants);
-      if (childTaskIds.length > 0) {
-        versionsToRestart.push({
-          versionId: childVersion.id,
-          taskIds: childTaskIds,
-        });
-      }
-    });
-
-    const totalFailedTasks = versionsToRestart.reduce(
-      (sum, v) => sum + v.taskIds.length,
-      0,
-    );
-
-    if (totalFailedTasks === 0) {
-      dispatchToast.warning("No failed tasks to restart.");
-      return;
-    }
-
-    sendEvent({
-      name: "Clicked restart failed tasks button",
-      abort: false,
-      "task.modified_count": totalFailedTasks,
-    });
-
-    restartVersions({
-      variables: {
-        versionId: patchId,
-        versionsToRestart,
-        abort: false,
-      },
-    });
-  };
 
   return (
     <>
@@ -159,17 +45,15 @@ export const RestartPatch: React.FC<RestartPatchProps> = ({
           isButton ? (
             <SplitButton
               data-cy="restart-version"
-              disabled={disabled || queryLoading || mutationLoading}
+              disabled={disabled}
               label="Restart"
               menuItems={[
-                <MenuItem
+                <RestartFailedTasks
                   key="restart-failed"
-                  data-cy="restart-failed-tasks"
-                  disabled={disabled || queryLoading || mutationLoading}
-                  onClick={handleRestartFailedTasks}
-                >
-                  Restart failed tasks
-                </MenuItem>,
+                  disabled={disabled}
+                  patchId={patchId}
+                  refetchQueries={refetchQueries}
+                />,
               ]}
               onClick={onClick}
               size="small"
